requirements.txt
altair==4.2.0
backports.zoneinfo==0.2.1
blinker==1.5
brotlipy==0.7.0
cachetools==5.2.0
click==8.1.3
commonmark==0.9.1
contourpy==1.0.5
cycler==0.11.0
fonttools==4.37.4
gitdb==4.0.9
GitPython==3.1.27
importlib-metadata==5.0.0
joblib==1.2.0
kiwisolver==1.4.4
matplotlib==3.6.0
nltk==3.7
numpy==1.23.3
pandas==1.5.0
pdfminer.six==20220524
pdfplumber==0.7.5
Pillow==9.2.0
plotly==5.10.0
protobuf==3.20.3
pyarrow==9.0.0
pydeck==0.8.0b3
Pympler==1.0.1
pytz-deprecation-shim==0.1.0.post0
regex==2022.9.13
rich==12.6.0
semver==2.13.0
smmap==5.0.0
streamlit==1.13.0
tenacity==8.1.0
toml==0.10.2
toolz==0.12.0
tqdm==4.64.1
tzdata==2022.4
tzlocal==4.2
validators==0.20.0
Wand==0.6.10
webencodings==0.5.1
wordcloud==1.8.2.2

